data re-organization and whitespace cleanup.
For ease of updates and additions, the players and vs data are now
under the src/ directory. The old utils are now repurposed to generate
the _players and _vs collections (see src/bin/generate).

New events can be added under src/data/<collection>/<event> and the
generate script will automatically merge the data.